rate tester updates
